requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
streamlit==1.30.0
moviepy>=2.0.0
yt-dlp>=2023.12.30
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.26.0
pytest>=7.4.0
pytest-mock>=3.12.0
gunicorn>=21.2.0
